The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Daily and Cumulative Cases.

Daily new confirmed, recovered, death and active cases with their cumulative cases from 2020-01-22 to 2020-10-07.

df_daily

Format

A data frame with 9 variables.

date

Date in YYYY-MM-DD format. The date from 2020-01-22 to 2020-10-07.

confirmed

Daily new confirmed cases.

death

Daily new mortality cases.

recovered

Daily new recovered cases.

active

Daily new active cases, which could calculated by 'active cases = confirmed - deaths - recoveries'.

confirmed_cum

Cumulative confirmed cases from 2020-01-22 to 2020-10-07.

death_cum

Cumulative death cases from 2020-01-22 to 2020-10-07.

recovered_cum

Cumulative recovered cases from 2020-01-22 to 2020-10-07.

active_cum

Cumulative active cases from 2020-01-22 to 2020-10-07, which could calculated by 'cumulative active cases = cumulative confirmed - cumulative deaths - cumulative recoveries'

Source

Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus website.

Details

The dataset contains the daily summary of Coronavirus cases (confirmed, death, and recovered), in global.